Remove modal and update content in Skills and Projects
The modal functionality in the Skills component has been removed. The information in the Content.js file for both 'Skills' and 'Projects' has been updated to provide a simpler and clearer presentation. The "READ MORE" button in the Projects.jsx file has been replaced with a paragraph detailing the project.